[domain] move KeyModified into module
