+++ domain/src/main/scala/net/kemitix/s3thorp/domain/KeyModified.scala
@@ -1,6 +1,4 @@
-package net.kemitix.s3thorp
-
-import net.kemitix.s3thorp.domain.{LastModified, RemoteKey}
+package net.kemitix.s3thorp.domain
 
 final case class KeyModified(key: RemoteKey,
                              modified: LastModified) {
