+++ src/main/scala/net/kemitix/s3thorp/awssdk/S3ObjectsByHash.scala
@@ -1,7 +1,6 @@
 package net.kemitix.s3thorp.awssdk
 
-import net.kemitix.s3thorp.domain.{LastModified, MD5Hash, RemoteKey}
-import net.kemitix.s3thorp.KeyModified
+import net.kemitix.s3thorp.domain.{KeyModified, LastModified, MD5Hash, RemoteKey}
 import software.amazon.awssdk.services.s3.model.S3Object
 
 trait S3ObjectsByHash {
