+++ src/main/scala/net/kemitix/s3thorp/awssdk/S3ObjectsData.scala
@@ -1,7 +1,7 @@
 package net.kemitix.s3thorp.awssdk
 
-import net.kemitix.s3thorp.domain.{MD5Hash, RemoteKey}
-import net.kemitix.s3thorp.{HashModified, KeyModified}
+import net.kemitix.s3thorp.domain.{KeyModified, MD5Hash, RemoteKey}
+import net.kemitix.s3thorp.HashModified
 
 /**
   * A list of objects and their MD5 hash values.
