+++ src/test/scala/net/kemitix/s3thorp/S3MetaDataEnricherSuite.scala
@@ -4,7 +4,7 @@ import java.io.File
 import java.time.Instant
 
 import net.kemitix.s3thorp.awssdk.S3ObjectsData
-import net.kemitix.s3thorp.domain.{Bucket, LastModified, LocalFile, MD5Hash, RemoteKey, RemoteMetaData, S3MetaData}
+import net.kemitix.s3thorp.domain.{Bucket, KeyModified, LastModified, LocalFile, MD5Hash, RemoteKey, RemoteMetaData, S3MetaData}
 
 class S3MetaDataEnricherSuite
   extends UnitTest
